Allow for relative imports to be skipped, even if file doesn't exist.
See rollup/rollup#532 for the explanation of the problem and the
use case for skipping relative imports for files that don't exist.
This pull request, in conjunction with rollup/rollup#572, resolves the
issue.

The revving of the rollup version in the package.json is because the
new test requires rollup@^0.25.7, and npm's caret semver behavior
[does not install 0.25.x when a dependency is listed at ^0.22.0](https://docs.npmjs.com/misc/semver#caret-ranges-123-025-004)